requirements.txt
aiogram==2.13
python-dotenv==0.20.0